requirements.txt
Django==4.2.9
djangorestframework==3.14.0
drf-spectacular==0.27.0
PyPDF2==3.0.1
gunicorn==21.2.0
pytz==2023.3
whitenoise==6.5.0